chore: setup home route
