+++ packages/dashboard/package.json
@@ -14,6 +14,7 @@
     "postcss": "^8.4.27",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
     "tailwindcss": "^3.3.3"
   },
   "devDependencies": {
