+++ packages/dashboard/src/App.tsx
@@ -1,33 +1,18 @@
-import { useState } from 'react'
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./App.css";
 
-const projects = [
-  {
-    title: 'Card 1',
-  },
-  {
-    title: 'Card 2',
-  },
-];
+import Home from "./Home";
 
 function App() {
-  const [count, setCount] = useState(0)
+  
 
   return (
-    <>
-     <nav className="bg-indigo-600 p-4 py-6 flex justify-between items-center">
-        <h1 className="text-white text-xl font-semibold">Michi</h1>
-        <button className="bg-red-500 text-white px-4 py-2 rounded-lg">Sign Out</button>
-      </nav>
-      <div className="container mx-auto mt-4 px-4 flex flex-col">
-        {projects.map((project, index) => (
-          <div key={index} className="bg-white text-black rounded-lg shadow-md p-6 mb-4">
-            <h2 className="text-xl font-semibold mb-2">{project.title}</h2>
-          </div>
-        ))}
-      </div>
-    </>
-  )
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home  />} />
+      </Routes>
+    </BrowserRouter>
+  );  
 }
 
-export default App
+export default App;
